% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picosat-core.R
\name{pico_solve_satire}
\alias{pico_solve_satire}
\title{Solve simple using 'satire'}
\usage{
pico_solve_satire(sat, max_solutions = 1, remove = "^dummy")
}
\arguments{
\item{sat}{SAT problem definition as created by \code{satire::sat_new()} or
\code{satire::read_dimacs()}}

\item{max_solutions}{limit number of solutions}

\item{remove}{regular expression for variables to remove when blocking solutions
and assembling values to return. Default: "^dummy" will block all
variables starting with the word "dummy" (as this is how the 'satire' 
package automatically creates dummy variables.)
If NULL no variables will be removed.}
}
\value{
Return NULL is problem is unsatisfiable.  Otherwise return a data.frame
        of solutions where each column represents a named variable in the 
        problem and each row is a solution.
}
\description{
Solve simple using 'satire'
}
\examples{
sat <- satire::sat_new()
satire::sat_add_exprs(sat, "a -> (b & c)")
pico_solve_satire(sat, max_solutions = 10)
satire::sat_solve_naive(sat)
satire::sat_solve_dpll(sat, max_solutions = 10)
}
