% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picosatire-package.R
\docType{package}
\name{picosatire-package}
\alias{picosatire}
\alias{picosatire-package}
\title{picosatire: SAT Solver using 'PicoSAT'}
\description{
A solver for Boolean satisfiability problems ('SAT') using the 'PicoSAT' library. For this solver, 'SAT' problems can be formulated in two ways; using integer literals, or using logical expressions specified with the 'satire' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coolbutuseless/picosatire}
  \item Report bugs at \url{https://github.com/coolbutuseless/picosatire/issues}
}

}
\author{
\strong{Maintainer}: Mike Cheng \email{mikefc@coolbutuseless.com} [copyright holder]

Other contributors:
\itemize{
  \item Dirk Schumacher \email{mail@dirk-schumacher.net} (Fixes for original picosat.c) [contributor]
  \item Armin Biere (Author and copyright holder of included PicoSAT code) [copyright holder]
}

}
\keyword{internal}
